+++ src/Telegram.Bot/TelegramBotClient.cs
@@ -783,6 +783,23 @@ namespace Telegram.Bot
             }, cancellationToken);
 
         /// <inheritdoc />
+        public Task<Message> SendDiceAsync(
+            ChatId chatId,
+            bool disableNotification = default,
+            int replyToMessageId = default,
+            IReplyMarkup replyMarkup = default,
+            CancellationToken cancellationToken = default) =>
+            MakeRequestAsync(
+                new SendDiceRequest(chatId)
+                {
+                    DisableNotification = disableNotification,
+                    ReplyToMessageId = replyToMessageId,
+                    ReplyMarkup = replyMarkup
+                },
+                cancellationToken
+            );
+
+        /// <inheritdoc />
         public Task SendChatActionAsync(
             ChatId chatId,
             ChatAction chatAction,
@@ -1001,6 +1018,16 @@ namespace Telegram.Bot
             CancellationToken cancellationToken = default
         ) =>
             MakeRequestAsync(new SetChatPermissionsRequest(chatId, permissions), cancellationToken);
+
+        /// <inheritdoc />
+        public Task<BotCommand[]> GetMyCommandsAsync(CancellationToken cancellationToken = default) =>
+            MakeRequestAsync(new GetMyCommandsRequest(), cancellationToken);
+
+        /// <inheritdoc />
+        public Task SetMyCommandsAsync(
+            IEnumerable<BotCommand> commands,
+            CancellationToken cancellationToken = default) =>
+            MakeRequestAsync(new SetMyCommandsRequest(commands), cancellationToken);
 
         /// <inheritdoc />
         public Task<Message> StopMessageLiveLocationAsync(
@@ -1235,7 +1262,9 @@ namespace Telegram.Bot
             bool disableNotification = default,
             int replyToMessageId = default,
             InlineKeyboardMarkup replyMarkup = default,
-            CancellationToken cancellationToken = default
+            CancellationToken cancellationToken = default,
+            bool sendPhoneNumberToProvider = default,
+            bool sendEmailToProvider = default
         ) =>
             MakeRequestAsync(new SendInvoiceRequest(
                 chatId,
@@ -1258,6 +1287,8 @@ namespace Telegram.Bot
                 NeedPhoneNumber = needPhoneNumber,
                 NeedEmail = needEmail,
                 NeedShippingAddress = needShippingAddress,
+                SendPhoneNumberToProvider = sendPhoneNumberToProvider,
+                SendEmailToProvider = sendEmailToProvider,
                 IsFlexible = isFlexible,
                 DisableNotification = disableNotification,
                 ReplyToMessageId = replyToMessageId,
@@ -1494,12 +1525,49 @@ namespace Telegram.Bot
             }, cancellationToken);
 
         /// <inheritdoc />
+        public Task CreateNewAnimatedStickerSetAsync(
+            int userId,
+            string name,
+            string title,
+            InputFileStream tgsSticker,
+            string emojis,
+            bool isMasks = default,
+            MaskPosition maskPosition = default,
+            CancellationToken cancellationToken = default) =>
+            MakeRequestAsync(
+                new CreateNewAnimatedStickerSetRequest(userId, name, title, tgsSticker, emojis)
+                {
+                    ContainsMasks = isMasks,
+                    MaskPosition = maskPosition
+                },
+                cancellationToken
+            );
+
+        /// <inheritdoc />
+        public Task AddAnimatedStickerToSetAsync(
+            int userId,
+            string name,
+            InputFileStream tgsSticker,
+            string emojis,
+            MaskPosition maskPosition = default,
+            CancellationToken cancellationToken = default) =>
+            MakeRequestAsync(
+                new AddAnimatedStickerToSetRequest(userId, name, tgsSticker, emojis)
+                {
+                    MaskPosition = maskPosition
+                },
+                cancellationToken
+            );
+
+        /// <inheritdoc />
         public Task SetStickerPositionInSetAsync(
             string sticker,
             int position,
-            CancellationToken cancellationToken = default
-        ) =>
-            MakeRequestAsync(new SetStickerPositionInSetRequest(sticker, position), cancellationToken);
+            CancellationToken cancellationToken = default) =>
+            MakeRequestAsync(
+                new SetStickerPositionInSetRequest(sticker, position),
+                cancellationToken
+            );
 
         /// <inheritdoc />
         public Task DeleteStickerFromSetAsync(
@@ -1507,6 +1575,17 @@ namespace Telegram.Bot
             CancellationToken cancellationToken = default
         ) =>
             MakeRequestAsync(new DeleteStickerFromSetRequest(sticker), cancellationToken);
+
+        /// <inheritdoc />
+        public Task SetStickerSetThumbAsync(
+            string name,
+            int userId,
+            InputOnlineFile thumb = default,
+            CancellationToken cancellationToken = default) =>
+            MakeRequestAsync(
+                new SetStickerSetThumbRequest(name, userId, thumb),
+                cancellationToken
+            );
 
         #endregion
     }