+++ test/Telegram.Bot.Tests.Integ/Framework/Constants.cs
@@ -56,6 +56,10 @@ namespace Telegram.Bot.Tests.Integ.Framework
 
             public const string FileDownload = "File Download";
 
+            public const string BotCommands = "Bot Commands";
+
+            public const string Dice = "Dice";
+
             public const string CallbackQuery = "Callback Query";
 
             public const string AlbumMessage = "Sending Album Messages";
@@ -300,6 +304,12 @@ namespace Telegram.Bot.Tests.Integ.Framework
             public const string SendPoll = "sendPoll";
 
             public const string StopPoll = "stopPoll";
+
+            public const string SetMyCommands = "setMyCommands";
+
+            public const string GetMyCommands = "getMyCommands";
+
+            public const string SendDice = "sendDice";
         }
     }
 }