+++ test/Telegram.Bot.Tests.Integ/Framework/XunitExtensions/XunitTestAssemblyRunnerWithAssemblyFixture.cs
@@ -11,7 +11,8 @@ namespace Telegram.Bot.Tests.Integ.Framework.XunitExtensions
 {
     public class XunitTestAssemblyRunnerWithAssemblyFixture : XunitTestAssemblyRunner
     {
-        readonly Dictionary<Type, object> _assemblyFixtureMappings = new Dictionary<Type, object>();
+        private readonly Dictionary<Type, object> _assemblyFixtureMappings =
+            new Dictionary<Type, object>();
 
         public XunitTestAssemblyRunnerWithAssemblyFixture(
             ITestAssembly testAssembly,
@@ -19,7 +20,12 @@ namespace Telegram.Bot.Tests.Integ.Framework.XunitExtensions
             IMessageSink diagnosticMessageSink,
             IMessageSink executionMessageSink,
             ITestFrameworkExecutionOptions executionOptions)
-            : base(testAssembly, testCases, diagnosticMessageSink, executionMessageSink, executionOptions)
+            : base(
+                testAssembly,
+                testCases,
+                diagnosticMessageSink,
+                executionMessageSink,
+                executionOptions)
         { }
 
         protected override async Task AfterTestAssemblyStartingAsync()
@@ -39,8 +45,40 @@ namespace Telegram.Bot.Tests.Integ.Framework.XunitExtensions
                 // Instantiate all the fixtures
                 foreach (var fixtureAttr in fixturesAttrs)
                 {
-                    var fixtureInstance = Activator.CreateInstance(fixtureAttr.FixtureType);
-                    _assemblyFixtureMappings[fixtureAttr.FixtureType] = fixtureInstance;
+                    var type = fixtureAttr.FixtureType;
+
+                    var ctors = type.GetConstructors();
+                    if (ctors.Length > 1)
+                    {
+                        throw new InvalidOperationException(
+                            "Assembly fixture can only have a single constructor"
+                        );
+                    }
+
+                    var ctor = ctors.Single();
+
+                    var parameters = ctor.GetParameters();
+                    if (parameters.Length > 1)
+                    {
+                        throw new InvalidOperationException(
+                            $"Assembly fixture constructor can only have a single parameter of type {nameof(IMessageSink)}"
+                        );
+                    }
+
+                    var parameter = parameters.SingleOrDefault();
+                    if (parameter?.ParameterType != typeof(IMessageSink))
+                    {
+                        throw new InvalidOperationException(
+                            $"Assembly fixture constructor can only have a parameter of type {nameof(IMessageSink)}"
+                        );
+                    }
+
+                    var fixture = ctor.Invoke(parameters.Length == 1
+                        ? new object[] { DiagnosticMessageSink }
+                        : new object[0]
+                    );
+
+                    _assemblyFixtureMappings[fixtureAttr.FixtureType] = fixture;
                 }
             });
         }