+++ CHANGELOG.md
@@ -19,7 +19,43 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 -->
 
-## [Unreleased]
+## [15.5.1] - 2020-04-02
+
+### Fixed
+- Implementation of `ITelegramBotClient.CreateNewAnimatedSticketSetAsync`
+
+## [15.5.0] - 2020-04-02
+
+### Added
+- Requests
+    - `GetMyCommandsRequest`
+    - `SetMyCommandsRequest`
+    - `CreateNewAnimatedStickerSetRequest`
+    - `AddNewAnimatedStickerToSetRequest`
+    - `SendDiceRequest`
+    - `SetStickerSetThumbRequest`
+- Methods:
+    - `ITelegramBotClient.SendDiceAsync`
+    - `ITelegramBotClient.CreateNewAnimatedSticketSetAsync`
+    - `ITelegramBotClient.AddNewAnimatedSticketToSetAsync`
+    - `ITelegramBotClient.SetStickerSetThumbAsync`
+    - `ITelegramBotClient.GetMyCommandsAsync`
+    - `ITelegramBotClient.SetMyCommandsAsync`
+- Type `Dice`
+- Type `BotCommand`
+- Enum member `MessageType.Dice`
+- Property `Message.Dice`
+- Property `SticketSet.Thumb`
+
+## [15.4.0] - 2020-02-22
+
+### Added
+- Property `SendInvoiceRequest.SendPhoneNumberToProvider`
+- Property `SendInvoiceRequest.SendEmailToProvider `
+- Optional parameter `sendPhoneNumberToProvider` to method `ITelegramBotClient.SendInvoiceAsync`
+- Optional parameter `sendEmailToProvider` to method `ITelegramBotClient.SendInvoiceAsync`
+
+## [15.3.0] - 2020-01-31
 
 ### Added
 
@@ -66,7 +102,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Type `ContactRequestException` is made obsolete due to Telegram changed it's error message
 
 ### Fixed
-- A bug that prevented requests to be serialized when default `JsonSerializerSettings` were set 
+- A bug that prevented requests to be serialized when default `JsonSerializerSettings` were set
 
 ## [15.2.0] - 2020-01-03